backend/requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
imagehash
Pillow
cloudinary
httpx
python-multipart
APScheduler==3.10.4